fix(Exercise Page): Fetch and display the exercise data
The page now draws an ExerciseCard component which displays the basic info of the exercise and uses two NewsBodyComponents to render the Markdown elements
